refactor(server): merge GameStateMachine into GameSession
- Removed GameStateMachine and merged its functionality into GameSession, as GameSession was just a middleman before.
- Removed GameStateMachine, GameStateMachineTest, and GameCommand, as they are no longer needed.
- Renamed the field gameState of GameSession to state and updated all classes that use it accordingly.
- Merged GameStateMachine tests into GameSession tests.
