--- server/src/main/kotlin/at/aau/kuhhandel/server/service/GameCommand.kt
@@ -1,50 +1,0 @@
-package at.aau.kuhhandel.server.service
-
-/**
- * Server-side commands that move a game from one deterministic state to the next.
- */
-sealed interface GameCommand {
-    data class AddPlayer(
-        val playerId: String,
-        val playerName: String,
-    ) : GameCommand
-
-    data class RemovePlayer(
-        val playerId: String,
-    ) : GameCommand
-
-    data object StartGame : GameCommand
-
-    data object RevealCard : GameCommand
-
-    data object ChooseAuction : GameCommand
-
-    data class PlaceBid(
-        val bidderId: String,
-        val amount: Int,
-    ) : GameCommand
-
-    data object CloseAuction : GameCommand
-
-    data class ResolveAuction(
-        val auctioneerBuysCard: Boolean,
-    ) : GameCommand
-
-    data class ChooseTrade(
-        val challengedPlayerId: String,
-        val animalType: at.aau.kuhhandel.shared.enums.AnimalType,
-        val offeredMoneyCardIds: List<String> = emptyList(),
-    ) : GameCommand
-
-    data class OfferTrade(
-        val offeredMoneyCardIds: List<String>,
-    ) : GameCommand
-
-    data class RespondToTrade(
-        val respondingPlayerId: String,
-        val acceptsOffer: Boolean,
-        val counterOfferedMoneyCardIds: List<String> = emptyList(),
-    ) : GameCommand
-
-    data object FinishRound : GameCommand
-}
